% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{enable_fast_read.FCS}
\alias{enable_fast_read.FCS}
\title{Enable the optimized flowCore overload}
\usage{
enable_fast_read.FCS(quiet = FALSE)
}
\arguments{
\item{quiet}{Logical. If TRUE, suppresses the startup message.}
}
\description{
Enable the optimized flowCore overload
}
