% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcs_hexdump.R
\name{fcs_hex_dump}
\alias{fcs_hex_dump}
\title{Display the hexadecimal canonical view of a file part}
\usage{
fcs_hex_dump(filename, start = 0, nbytes = 4096, replaceChar = " ")
}
\arguments{
\item{filename}{FCS file path as a string.}

\item{start}{byte position; 0 means the first byte. If -1, target TEXT of
FCS.}

\item{nbytes}{number of bytes to display.}

\item{replaceChar}{character that are not ASCII and could not be displayed
are replaced by the is character.}
}
\value{
none.
}
\description{
Display the hexadecimal canonical view of a file part. The view consists in
three parts. On each line, the first is the position, next the hexadecimal
view, finally the character/ASCII view.
}
\examples{
library(flowCore)
fname <- system.file("extdata", "0877408774.B08", package = "flowCore")
fcs_hex_dump(fname)
\dontrun{
writeChar(fcs_hex_dump(fname, nbytes = 4800), con = "hexdump.txt")
}

}
