% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcs_simplify.R
\name{fcs_simplify}
\alias{fcs_simplify}
\title{Simplify a FCS file}
\usage{
fcs_simplify(
  filename,
  mode = list("removeFlowCoreKeywords"),
  emptyValue = TRUE
)
}
\arguments{
\item{filename}{a file name.}

\item{mode}{a list of actions; currently only "removeFlowCoreKeywords".}

\item{emptyValue}{see read.FCS documentation.}
}
\value{
None.
}
\description{
The function \code{fcs_simplify} aims to keep FCS file as simple as possible.
The currently defined action removes flowCore_ keywords from the header of
the given FCS file. It works in place, i.e. writing the changed header in the
file on disk. Please make a copy of the original file before using.
}
\examples{
library(flowCore)
fname <- system.file("extdata", "0877408774.B08", package = "flowCore")
fcs_simplify(fname)
}
