% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcu_read_fcs.R
\name{fcu_read_fcs}
\alias{fcu_read_fcs}
\title{fcu_read_fcs}
\usage{
fcu_read_fcs(
  filename,
  which.lines = NULL,
  transformation = FALSE,
  truncate_max_range = FALSE,
  method = c("Rcpp", "default", "readBin"),
  ...
)
}
\arguments{
\item{filename}{file path to the FCS file}

\item{which.lines}{see flowCore::read.FCS}

\item{transformation}{see flowCore::read.FCS}

\item{truncate_max_range}{see flowCore::read.FCS}

\item{method}{one of "default", "readBin", "RCpp"}

\item{...}{arguments passed to flowCore}
}
\value{
a flowFrame
}
\description{
This function speeds up the reading of the DATA segment of an FCS file when
the DATA are float. Otherwise, it reverts to flowCore read.FCS(). No
operation is applied after the data were read: no transform, no truncate,
etc.
}
