% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_utils.R
\name{ff_clean_keyword}
\alias{ff_clean_keyword}
\title{Clean keywords in a flowFrame}
\usage{
ff_clean_keyword(
  ff,
  remove = c("^flowCore_", "transformation", "note"),
  keep = NULL
)
}
\arguments{
\item{ff}{a flowFrame}

\item{remove}{a vector of patterns to search and remove}

\item{keep}{unused yet}
}
\value{
the cleaned flowFrame
}
\description{
By default, it removes flowCore_ keywords as well as "transformation".
}
\examples{
library(flowCore)
fcsFile <- system.file("extdata", "0877408774.B08", package = "flowCore")
## read file and linearize values
samp <- read.FCS(fcsFile, transformation = "linearize")
names(keyword(samp))
## Remove flowCore keywords
samp <- ff_clean_keyword(samp)
names(keyword(samp))
}
