% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{restore_original_read.FCS}
\alias{restore_original_read.FCS}
\title{Disable the optimized flowCore overload}
\usage{
restore_original_read.FCS(quiet = FALSE)
}
\arguments{
\item{quiet}{Logical. If TRUE, suppresses the message.}
}
\description{
Disable the optimized flowCore overload
}
